extract shallow currenRoute from router
